requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.5
urllib3==2.0.7
